adds custom events
