+++ src/constants/events.ts
@@ -1,0 +1,6 @@
+export const PROVIDER_EVENT = 'provider-event';
+export const CHAIN_CHANGED = 'chainChanged';
+export const ACCOUNT_CHANGED = 'accountChanged';
+export const CONNECT = 'connect';
+export const DISCONNECT = 'disconnect';
+export const UNSUPPORTED_CHAINS_IDS_MESSAGE = 'Switch to a supported chain:';
