+++ src/hooks/useProviderListeners.ts
@@ -1,6 +1,8 @@
+import { CHAIN_CHANGED, CONNECT, ACCOUNT_CHANGED, DISCONNECT } from './../constants/events';
 import { DWPConfig } from '../types';
 import { useState, useEffect, useMemo } from 'react';
 import Web3Modal from 'web3modal';
+import { emitUnsupportedChainEvent } from '../helpers';
 
 export const useProviderListeners = (
   web3Modal: Web3Modal,
@@ -20,20 +22,25 @@ export const useProviderListeners = (
 
   useEffect(() => {
     // subscribe to connect events
-    web3Modal.on('connect', async _modalProvider => {
+    web3Modal.on(CONNECT, async _modalProvider => {
       // check that connected chain is supported
       if (config.supportedChains.includes(parseInt(_modalProvider.chainId))) {
         setModalProvider(_modalProvider);
+
+        // check that fallback id is supported (prevents endless loop)
       } else if (config.supportedChains.includes(parseInt(config.fallbackChainId))) {
+        emitUnsupportedChainEvent(CONNECT, config.supportedChains.join(', '));
         await connectDefaultProvider();
         setModalProvider(null);
+
+        // connect connect to provider, error in config
       } else {
-        // @todo event cannot connect
+        emitUnsupportedChainEvent(CONNECT, config.supportedChains.join(', '));
         setModalProvider(null);
       }
     });
     return () => {
-      web3Modal.off('connect');
+      web3Modal.off(CONNECT);
     };
   }, [web3Modal, config, connectDefaultProvider]);
 
@@ -41,6 +48,7 @@ export const useProviderListeners = (
     const chainChangedCallback = async (chainId: string) => {
       if (!config.supportedChains.includes(parseInt(chainId))) {
         // switch to a default provider
+        emitUnsupportedChainEvent(CHAIN_CHANGED, config.supportedChains.join(', '));
         connectDefaultProvider();
       } else {
         await web3Modal.connect();
@@ -64,13 +72,13 @@ export const useProviderListeners = (
     if (!modalProvider) return;
 
     // subscribe to chain events
-    modalProvider.on('chainChanged', chainChangedCallback);
+    modalProvider.on(CHAIN_CHANGED, chainChangedCallback);
 
     // subscribe to account change events
-    modalProvider.on('accountsChanged', accountsChangedCallback);
+    modalProvider.on(ACCOUNT_CHANGED, accountsChangedCallback);
 
     // subscribe to provider disconnection
-    modalProvider.on('disconnect', disconnectCallback);
+    modalProvider.on(DISCONNECT, disconnectCallback);
   }, [modalProvider, web3Modal, config, connectDefaultProvider]);
 
   return modalProvider;
