+++ src/index.ts
@@ -1,4 +1,5 @@
 export { parseChainsIds } from './utils/chains';
 export { Web3Provider } from './Web3Provider';
 export { useWeb3Provider } from './hooks/useWeb3Provider';
-export type { InitialState } from './types';
+export type { WalletProviderEvent } from './types';
+export { PROVIDER_EVENT } from './constants';
